Prevent time prediction updating twice during the same tick
This was causing the timer to update twice, making it a little "unpredictable" by 0.5 seconds. This is because the AREA_POS_UPDATED is updated twice; First for alliance, then horde.
Now we just do a delay check to only update once whether both sides own points, or just one side (when it's only fired once)