Add health syncing mod for Ashenvale
The changes to the health tracker are a bit messy, but it works.
Would be nicer to refactor it into a proper object some time.

Also indirectly improves other mods using the health tracker (Alterac)
because the list is now sorted by config and shows bosses prior to first
sync.